refactor: simplify drop_index() with NameIdValueApi
